Moved 'MathExtension" to 'markdown_extensions.py'
